group DTD validation tests by area
