+++ parser_test.go
@@ -1665,347 +1665,6 @@ func TestParseExternalEntityValidEncoding(t *testing.T) {
 	require.NoError(t, err, "well-formed UTF-16 external entity must load")
 }
 
-func TestValidateDTD(t *testing.T) {
-	t.Parallel()
-
-	t.Run("required attribute missing", func(t *testing.T) {
-		t.Parallel()
-
-		// #REQUIRED attribute missing -> validation error
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc EMPTY>
-  <!ATTLIST doc id ID #REQUIRED>
-]>
-<doc/>`
-
-		collector := helium.NewErrorCollector(t.Context(), helium.ErrorLevelNone)
-		p := helium.NewParser().ValidateDTD(true).ErrorHandler(collector)
-		doc, err := p.Parse(t.Context(), []byte(input))
-
-		require.Error(t, err, "missing #REQUIRED attribute should fail validation")
-		require.NotNil(t, doc, "document should still be returned with validation error")
-		require.ErrorIs(t, err, helium.ErrDTDValidationFailed)
-		require.True(t, containsError(collector.Errors(), "required"))
-	})
-
-	t.Run("required attribute present", func(t *testing.T) {
-		t.Parallel()
-
-		// #REQUIRED attribute present -> no error
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc EMPTY>
-  <!ATTLIST doc id ID #REQUIRED>
-]>
-<doc id="x1"/>`
-
-		p := helium.NewParser().ValidateDTD(true)
-		_, err := p.Parse(t.Context(), []byte(input))
-		require.NoError(t, err)
-	})
-
-	t.Run("fixed mismatch", func(t *testing.T) {
-		t.Parallel()
-
-		// #FIXED attribute with wrong value -> validation error
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc EMPTY>
-  <!ATTLIST doc version CDATA #FIXED "1.0">
-]>
-<doc version="2.0"/>`
-
-		collector := helium.NewErrorCollector(t.Context(), helium.ErrorLevelNone)
-		p := helium.NewParser().ValidateDTD(true).ErrorHandler(collector)
-		_, err := p.Parse(t.Context(), []byte(input))
-
-		require.ErrorIs(t, err, helium.ErrDTDValidationFailed)
-		require.True(t, containsError(collector.Errors(), "must be"))
-	})
-
-	t.Run("fixed correct", func(t *testing.T) {
-		t.Parallel()
-
-		// #FIXED attribute with correct value -> no error
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc EMPTY>
-  <!ATTLIST doc version CDATA #FIXED "1.0">
-]>
-<doc version="1.0"/>`
-
-		p := helium.NewParser().ValidateDTD(true)
-		_, err := p.Parse(t.Context(), []byte(input))
-		require.NoError(t, err)
-	})
-
-	t.Run("empty element with content", func(t *testing.T) {
-		t.Parallel()
-
-		// EMPTY element with content -> validation error
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (child)>
-  <!ELEMENT child EMPTY>
-]>
-<doc><child>text</child></doc>`
-
-		collector := helium.NewErrorCollector(t.Context(), helium.ErrorLevelNone)
-		p := helium.NewParser().ValidateDTD(true).ErrorHandler(collector)
-		_, err := p.Parse(t.Context(), []byte(input))
-
-		require.ErrorIs(t, err, helium.ErrDTDValidationFailed)
-		require.True(t, containsError(collector.Errors(), "EMPTY"))
-	})
-
-	t.Run("element content valid", func(t *testing.T) {
-		t.Parallel()
-
-		// Element content model (a, b) with correct content -> no error
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (a, b)>
-  <!ELEMENT a (#PCDATA)>
-  <!ELEMENT b (#PCDATA)>
-]>
-<doc><a>hello</a><b>world</b></doc>`
-
-		p := helium.NewParser().ValidateDTD(true)
-		_, err := p.Parse(t.Context(), []byte(input))
-		require.NoError(t, err)
-	})
-
-	t.Run("element content mismatch", func(t *testing.T) {
-		t.Parallel()
-
-		// Element content model (a, b) with (b, a) -> validation error
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (a, b)>
-  <!ELEMENT a (#PCDATA)>
-  <!ELEMENT b (#PCDATA)>
-]>
-<doc><b>world</b><a>hello</a></doc>`
-
-		p := helium.NewParser().ValidateDTD(true)
-		_, err := p.Parse(t.Context(), []byte(input))
-		require.Error(t, err, "wrong element order should fail content model")
-	})
-
-	t.Run("mixed content valid", func(t *testing.T) {
-		t.Parallel()
-
-		// Mixed content (#PCDATA | a)* -- text and <a> are allowed
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (#PCDATA | a)*>
-  <!ELEMENT a (#PCDATA)>
-]>
-<doc>hello <a>world</a> end</doc>`
-
-		p := helium.NewParser().ValidateDTD(true)
-		_, err := p.Parse(t.Context(), []byte(input))
-		require.NoError(t, err)
-	})
-
-	t.Run("mixed content bad child", func(t *testing.T) {
-		t.Parallel()
-
-		// Mixed content (#PCDATA | a)* -- <b> is NOT allowed
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (#PCDATA | a)*>
-  <!ELEMENT a (#PCDATA)>
-  <!ELEMENT b (#PCDATA)>
-]>
-<doc>hello <b>world</b></doc>`
-
-		p := helium.NewParser().ValidateDTD(true)
-		_, err := p.Parse(t.Context(), []byte(input))
-		require.Error(t, err, "<b> not allowed in mixed content (a)")
-	})
-
-	t.Run("no flag skips validation", func(t *testing.T) {
-		t.Parallel()
-
-		// Same invalid document but WITHOUT ValidateDTD -> should pass
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc EMPTY>
-  <!ATTLIST doc id ID #REQUIRED>
-]>
-<doc/>`
-
-		p := helium.NewParser()
-		// Don't set ValidateDTD
-		_, err := p.Parse(t.Context(), []byte(input))
-		require.NoError(t, err, "without ValidateDTD, validation should not run")
-	})
-
-	t.Run("choice content", func(t *testing.T) {
-		t.Parallel()
-
-		// Choice content model (a | b) with <a> -> valid
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (a | b)>
-  <!ELEMENT a (#PCDATA)>
-  <!ELEMENT b (#PCDATA)>
-]>
-<doc><a>hello</a></doc>`
-
-		p := helium.NewParser().ValidateDTD(true)
-		_, err := p.Parse(t.Context(), []byte(input))
-		require.NoError(t, err)
-	})
-
-	t.Run("repeat content", func(t *testing.T) {
-		t.Parallel()
-
-		// Repetition content model (a)+ with multiple <a> -> valid
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (a)+>
-  <!ELEMENT a (#PCDATA)>
-]>
-<doc><a>1</a><a>2</a><a>3</a></doc>`
-
-		p := helium.NewParser().ValidateDTD(true)
-		_, err := p.Parse(t.Context(), []byte(input))
-		require.NoError(t, err)
-	})
-
-	t.Run("repeat content empty", func(t *testing.T) {
-		t.Parallel()
-
-		// Repetition content model (a)+ with zero <a> -> invalid
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (a)+>
-  <!ELEMENT a (#PCDATA)>
-]>
-<doc></doc>`
-
-		p := helium.NewParser().ValidateDTD(true)
-		_, err := p.Parse(t.Context(), []byte(input))
-		require.Error(t, err, "(a)+ requires at least one <a>")
-	})
-
-	t.Run("ID unique", func(t *testing.T) {
-		t.Parallel()
-
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (item, item)>
-  <!ELEMENT item EMPTY>
-  <!ATTLIST item id ID #REQUIRED>
-]>
-<doc><item id="a"/><item id="b"/></doc>`
-
-		p := helium.NewParser().ValidateDTD(true)
-		_, err := p.Parse(t.Context(), []byte(input))
-		require.NoError(t, err)
-	})
-
-	t.Run("ID duplicate", func(t *testing.T) {
-		t.Parallel()
-
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (item, item)>
-  <!ELEMENT item EMPTY>
-  <!ATTLIST item id ID #REQUIRED>
-]>
-<doc><item id="a"/><item id="a"/></doc>`
-
-		collector := helium.NewErrorCollector(t.Context(), helium.ErrorLevelNone)
-		p := helium.NewParser().ValidateDTD(true).ErrorHandler(collector)
-		_, err := p.Parse(t.Context(), []byte(input))
-
-		require.ErrorIs(t, err, helium.ErrDTDValidationFailed)
-		require.True(t, containsError(collector.Errors(), "duplicate ID"))
-	})
-
-	t.Run("IDRef valid", func(t *testing.T) {
-		t.Parallel()
-
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (item, ref)>
-  <!ELEMENT item EMPTY>
-  <!ELEMENT ref EMPTY>
-  <!ATTLIST item id ID #REQUIRED>
-  <!ATTLIST ref target IDREF #REQUIRED>
-]>
-<doc><item id="x"/><ref target="x"/></doc>`
-
-		p := helium.NewParser().ValidateDTD(true)
-		_, err := p.Parse(t.Context(), []byte(input))
-		require.NoError(t, err)
-	})
-
-	t.Run("IDRef missing", func(t *testing.T) {
-		t.Parallel()
-
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (item, ref)>
-  <!ELEMENT item EMPTY>
-  <!ELEMENT ref EMPTY>
-  <!ATTLIST item id ID #REQUIRED>
-  <!ATTLIST ref target IDREF #REQUIRED>
-]>
-<doc><item id="x"/><ref target="y"/></doc>`
-
-		collector := helium.NewErrorCollector(t.Context(), helium.ErrorLevelNone)
-		p := helium.NewParser().ValidateDTD(true).ErrorHandler(collector)
-		_, err := p.Parse(t.Context(), []byte(input))
-
-		require.ErrorIs(t, err, helium.ErrDTDValidationFailed)
-		require.True(t, containsError(collector.Errors(), "unknown ID"))
-	})
-
-	t.Run("IDRefs valid", func(t *testing.T) {
-		t.Parallel()
-
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (item, item, refs)>
-  <!ELEMENT item EMPTY>
-  <!ELEMENT refs EMPTY>
-  <!ATTLIST item id ID #REQUIRED>
-  <!ATTLIST refs targets IDREFS #REQUIRED>
-]>
-<doc><item id="a"/><item id="b"/><refs targets="a b"/></doc>`
-
-		p := helium.NewParser().ValidateDTD(true)
-		_, err := p.Parse(t.Context(), []byte(input))
-		require.NoError(t, err)
-	})
-
-	t.Run("IDRefs missing", func(t *testing.T) {
-		t.Parallel()
-
-		const input = `<?xml version="1.0"?>
-<!DOCTYPE doc [
-  <!ELEMENT doc (item, refs)>
-  <!ELEMENT item EMPTY>
-  <!ELEMENT refs EMPTY>
-  <!ATTLIST item id ID #REQUIRED>
-  <!ATTLIST refs targets IDREFS #REQUIRED>
-]>
-<doc><item id="a"/><refs targets="a z"/></doc>`
-
-		collector := helium.NewErrorCollector(t.Context(), helium.ErrorLevelNone)
-		p := helium.NewParser().ValidateDTD(true).ErrorHandler(collector)
-		_, err := p.Parse(t.Context(), []byte(input))
-
-		require.ErrorIs(t, err, helium.ErrDTDValidationFailed)
-		require.True(t, containsError(collector.Errors(), "unknown ID"))
-	})
-}
-
 func TestDTDDuplicateEnumerationTokens(t *testing.T) {
 	t.Parallel()
